requirements.txt
numpy==1.19.1
tqdm==4.46.1
torch==1.4.0
matplotlib==3.2.2
scikit-learn==0.23.1
scikit-image==0.17.2
torchvision==0.5.0
pandas==1.1.0
xarray==0.16.0
opencv-python==4.2.0.34
PyYAML==5.3.1
easydict==1.9
tensorboard==2.3.0
